app/requirements.txt
streamlit
tensorflow
pillow
numpy
pandas
matplotlib
seaborn
scikit-learn
tqdm
joblib